Move shapefile data source to non-spark package name, don't use package private methods
