+++ spark/spark-3.5/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,3 +1,3 @@
 org.apache.spark.sql.execution.datasources.parquet.GeoParquetFileFormat
 org.apache.spark.sql.execution.datasources.v2.geoparquet.metadata.GeoParquetMetadataDataSource
-org.apache.spark.sql.execution.datasources.shapefile.ShapefileDataSource
+org.apache.sedona.sql.datasources.shapefile.ShapefileDataSource
