+++ spark/spark-3.5/src/main/scala/org/apache/sedona/sql/datasources/shapefile/ShapefileDataSource.scala
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.spark.sql.execution.datasources.shapefile
+package org.apache.sedona.sql.datasources.shapefile
 
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.connector.catalog.Table
